Move global utils components to @threefold/common
