+++ packages/common/src/components/copy_input_wrapper.vue
