+++ packages/common/src/components/dynamic_tabs.vue
