+++ packages/common/src/global-components.d.ts
@@ -1,11 +1,19 @@
-import FormValidator from "./components/form_validator.vue";
-import InputValidator from "./components/input_validator.vue";
+import CopyInputWrapper from "../components/copy_input_wrapper.vue";
+import DTabs from "../components/dynamic_tabs.vue";
+import FormValidator from "../components/form_validator.vue";
+import InputTooltip from "../components/input_tooltip.vue";
+import InputValidator from "../components/input_validator.vue";
+import PasswordInputWrapper from "../components/password_input_wrapper.vue";
 import type { ValidatorsType } from "./utils";
 
 export declare module "@vue/runtime-core" {
   export interface GlobalComponents {
     InputValidator: typeof InputValidator;
     FormValidator: typeof FormValidator;
+    CopyInputWrapper: typeof CopyInputWrapper;
+    DTabs: typeof DTabs;
+    InputTooltip: typeof InputTooltip;
+    PasswordInputWrapper: typeof PasswordInputWrapper;
   }
 
   export interface ComponentCustomProperties {
