+++ packages/common/src/hooks/globals.ts
@@ -1,12 +1,20 @@
 import type { App, Component } from "vue";
 
+import CopyInputWrapper from "../components/copy_input_wrapper.vue";
+import DTabs from "../components/dynamic_tabs.vue";
 import FormValidator from "../components/form_validator.vue";
+import InputTooltip from "../components/input_tooltip.vue";
 import InputValidator from "../components/input_validator.vue";
+import PasswordInputWrapper from "../components/password_input_wrapper.vue";
 import * as validators from "../utils/validators";
 
 const GLOBAL_COMPONENTS: { [key: string]: Component } = {
   InputValidator,
   FormValidator,
+  PasswordInputWrapper,
+  CopyInputWrapper,
+  DTabs,
+  InputTooltip,
 };
 export function useGlobals(): any {
   return (app: App<any>) => {
