+++ packages/playground/src/global-components.d.ts
@@ -1,20 +1,12 @@
 import type { NetworkEnv } from "@threefold/grid_client";
 
-import CopyInputWrapper from "./components/copy_input_wrapper.vue";
-import DTabs from "./components/dynamic_tabs.vue";
-import InputTooltip from "./components/input_tooltip.vue";
-import PasswordInputWrapper from "./components/password_input_wrapper.vue";
 import ViewLayout from "./components/view_layout.vue";
 import WebletLayout from "./components/weblet_layout.vue";
 
 declare module "@vue/runtime-core" {
   export interface GlobalComponents {
-    PasswordInputWrapper: typeof PasswordInputWrapper;
     WebletLayout: typeof WebletLayout;
-    CopyInputWrapper: typeof CopyInputWrapper;
-    DTabs: typeof DTabs;
     ViewLayout: typeof ViewLayout;
-    InputTooltip: typeof InputTooltip;
   }
 }
